+++ package.json
@@ -8,6 +8,7 @@
   },
   "devDependencies": {
     "@types/express": "^4.17.18",
+    "@types/jsonwebtoken": "^9.0.3",
     "@types/mongoose": "^5.11.97"
   }
 }
