--- src/middlewares/auth.js
@@ -1,31 +1,0 @@
-const jwt = require('jsonwebtoken');
-require('dotenv').config();
-
-const auth = (req, res, next) => {
-    const authHeader = req.headers['authorization'];
-    const token = authHeader && authHeader.split(' ')[1];
-    if (!token) {
-      return res.status(401).json({ error: 'Unauthorized' });
-    }
-  
-    jwt.verify(token, process.env.SECRET_KEY, (err, user) => {
-      if (err) {
-        return res.status(403).json({ error: 'Forbidden' });
-      }
-      req.user = user;
-      next();
-    });
-}
-
-const isAdmin = (req, res, next) => {
-    if (req.user && req.user.isAdmin) {
-        next();
-    } else {
-        res.status(403).json({ error: 'Access forbidden' });
-    }
-}
-
-module.exports = {
-    auth,
-    isAdmin
-}
